Added monitoring module
